+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 1.0.7 - 
 ------------------
 
+* character encoding now defaults to ``"latin1"`` when
+  ``language_driver == 0``. It was previously ``"ascii"``, which
+  caused decoding errors with files that didn't in fact use ASCII.
+
 
 1.0.6 - 2014-08-02
 ------------------